Login - Sidebar / stats angular components makeover - removed 'popup'
* Measurement details for EOW Data points now displaying
