--- src/app/sidebar/popup.spec.ts
@@ -1,10 +1,0 @@
-import { Popup } from './popup';
-import * as chai from 'chai';
-
-const expect = chai.expect;
-
-describe('Popup', () => {
-  it('should create an instance', () => {
-    expect(new Popup(null, null)).is.not.undefined;  // tslint:disable-line
-  });
-});
