+++ src/app/sidebar/sidebar.component.html
@@ -15,7 +15,7 @@
 </div>
 <div class="panel-body" id="eow-dataPoint-information"> <!-- popup id -->
   <h5>Measurement details:<span class="close"></span></h5>
-  <div class="stats"><app-stats [stats]="stats" [pieChartPreparedData]="pieChartPreparedData" [pieChartParentSelector]="pieChartParentSelector"></app-stats></div>
+  <div class="stats"><app-stats [stats]="stats" [pieChartPreparedData]="pieChartPreparedData" [pieChartParentSelector]="parentSelector"></app-stats></div>
   <div class="content"></div>
 </div>
 <div class="panel-body" id="eow-timeline">
