+++ src/app/sidebar/sidebar.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, Inject, Input, OnInit} from '@angular/core';
 import {Stats, StatsService} from '../stats/stats.service';
 import SideBarService from './sidebar.service';
 import {EowBaseService} from '../eow-base-service';
@@ -8,6 +8,9 @@ import {Subject} from 'rxjs';
 import {PieChart} from '../charts/pie-chart';
 import {EowDataStruct, PieItem} from '../eow-data-struct';
 import Feature from 'ol/Feature';
+import {DOCUMENT} from '@angular/common';
+import colors from '../colors.json';
+import moment = require('moment-timezone/moment-timezone');
 
 @Component({
   selector: 'app-sidebar',
@@ -18,9 +21,9 @@ export class SidebarComponent extends EowBaseService implements OnInit {
   @Input() sideBarMessagingService: Subject<SideBarMessage>;
   stats: Stats;
   pieChartPreparedData: PieItem[];
-  pieChartParentSelector = 'div#eow-dataPoint-information';
+  parentSelector = 'div#eow-dataPoint-information';
 
-  constructor(private sideBarService: SideBarService, private statsService: StatsService, private log: Brolog) {
+  constructor(private sideBarService: SideBarService, private statsService: StatsService, @Inject(DOCUMENT) private htmlDocument: Document, private log: Brolog) {
     super();
     this.stats = new Stats();
   }
@@ -67,7 +70,91 @@ export class SidebarComponent extends EowBaseService implements OnInit {
   }
 
   private async showStats(features: Feature[]) {
-    this.stats = this.statsService.calculateStats(features);
-    this.pieChartPreparedData = await EowDataStruct.preparePieChartData(features);
+    const element = this.htmlDocument.querySelector(this.parentSelector);  // this.popup.getElement();
+    const content = element.querySelector('.content');
+    const stats = element.querySelector('.stats');
+    content.innerHTML = '';
+    element.classList.remove('active');
+
+    if (features.length) {
+      content.innerHTML = features.map(f => this.printDetails(f)).join('');
+      // stats.innerHTML = PieChart.fixForThisPieChart(printStats(calculateStats(features), this.userStore));
+      element.classList.add('active');
+      this.stats = this.statsService.calculateStats(features);
+      this.pieChartPreparedData = await EowDataStruct.preparePieChartData(features);
+      // const preparedFeatures = await EowDataStruct.preparePieChartData(features);
+      // PieChart.drawD3(preparedFeatures, '.pieChart', 8);
+    }
+
+    this.setupEventHandlers(this.parentSelector);
+  }
+
+  private setupEventHandlers(elementSelector) {
+    // Popup dialog close button
+    this.htmlDocument.querySelector(elementSelector).addEventListener('click', (event: Event) => {
+      const element = (event.target as HTMLElement);
+      if (element.matches('.close')) {
+        console.log(`close`);
+        this.sideBarMessagingService.next({action: 'close', message: 'eow-dataPoint-information'});
+        // send message to sidebar to close and reopen
+      }
+    });
+
+    const moreInfoButton = this.htmlDocument.querySelectorAll('.more-info-btn');
+    moreInfoButton.forEach(mib => {
+      mib.addEventListener('click', (event: Event) => {
+        const popupElement = mib.closest('.popup-item');
+        popupElement.classList.toggle('active');
+      });
+    });
+  }
+
+  private printDetails(feature) {
+    const properties = feature.getProperties();
+    const details = this.buildDetails(properties);
+    return `
+      <div class="popup-item">
+        <div class="metadata-row">
+          <div class="image">
+            <img src="${properties.image}" />
+          </div>
+          <div class="metadata">
+            <div class="fu-preview"  style="background:${colors[properties.fu_value]}"></div>
+            <div class="more-info-btn"></div>
+            <div class="thead"> FU value: ${properties.fu_value}</div>
+            <div class="thead"> Date: ${formatDate(properties.date_photo)}</div>
+            <div class="thead"> Device:  ${properties.device_model}</div>
+          </div>
+        </div>
+        <div class="raw-details">${details}</div>
+      </div>
+    `;
+  }
+
+  private buildDetails(properties) {
+    return `<table>
+              <tr><td class="thead">Lon, Lat</td><td colspan="5">${properties.lng}, ${properties.lat}</td></tr>
+              <tr><td class="thead">Device</td><td>${formatNull(properties.device_platform)}</td>
+                  <td class="thead">Model</td><td>${formatNull(properties.device_model)}</td></tr>
+              <tr><td class="thead">Viewing angle</td><td>${formatNull(properties.viewing_angle)}</td>
+                  <td class="thead">Cloud cover</td><td colspan="1">${formatNull(properties.p_cloud_cover)}</td></tr>
+              <tr><td class="thead">FU Value</td><td colspan="1">${formatNull(properties.fu_value)}</td>
+                    <td class="thead">Raining?</td><td>${formatNull(properties.rain)}</td></tr>
+              <tr><td class="thead">FU Observed</td><td colspan="1">${formatNull(properties.fu_observed)}</td>
+                    <td class="thead">See bottom?</td><td>${formatNull(properties.bottom)}</td></tr>
+              <tr><td class="thead">FU Processed</td><td colspan="1">${formatNull(properties.fu_processed)}</td>
+                     <td class="thead">Sechi depth</td><td colspan="1">${formatNull(properties.sd_depth)}</td></tr>
+              <tr><td class="thead">PH</td><td colspan="1">${formatNull(properties.p_ph)}</td>
+                  <td class="thead">Conductivity</td><td colspan="1">${formatNull(properties.p_conductivity)}</td></tr>
+        </table>`;
   }
 }
+
+const formatDate = (d) => {
+  return moment(d).format('MM/DD/YYYY hh:mm Z');
+};
+
+const formatNull = (d) => {
+  return d ? d : '';
+};
+
